requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
openai==0.27.0
pyyaml==5.4.1
pytest==6.2.5
